rename edge_echema to edge_schema
